refactor(compiler): split native project.rs into domain modules
Pure code move, no behavior change. The 2023-line project.rs (one giant
#[napi] impl Compiler block + free helpers) is split into a 317-line parent
(struct defs, from_config, mirror structs) plus project/{config,build_info,
files,css,recipes,codegen,introspect,transforms,support}.rs.
